+++ packages/backend-deployer/src/cdk_error_mapper.ts
@@ -5,13 +5,12 @@ import {
   AmplifyUserError,
 } from '@aws-amplify/platform-core';
 import { BackendDeployerOutputFormatter } from './types.js';
+import { EOL } from 'os';
 
 /**
  * Transforms CDK error messages to human readable ones
  */
 export class CdkErrorMapper {
-  private placeHolder = 'PLACEHOLDER';
-
   /**
    * Instantiate with a formatter that will be used for formatting CLI commands in error messages
    */
@@ -38,16 +37,19 @@ export class CdkErrorMapper {
 
       if (matchGroups && matchGroups.length > 1) {
         // If the contextual information can be used in the error message use it, else consider it as a downstream cause
-        if (
-          matchingError.humanReadableErrorMessage.includes(this.placeHolder)
-        ) {
-          matchingError.humanReadableErrorMessage =
-            matchingError.humanReadableErrorMessage.replace(
-              this.placeHolder,
-              matchGroups[1] // matching group instead of the matching string
-            );
-          // reset the stderr dump in the underlying error
-          underlyingError = undefined;
+        if (matchGroups.groups) {
+          for (const [key, value] of Object.entries(matchGroups.groups)) {
+            const placeHolder = `{${key}}`;
+            if (matchingError.humanReadableErrorMessage.includes(placeHolder)) {
+              matchingError.humanReadableErrorMessage =
+                matchingError.humanReadableErrorMessage.replace(
+                  placeHolder,
+                  value
+                );
+              // reset the stderr dump in the underlying error
+              underlyingError = undefined;
+            }
+          }
         } else {
           underlyingError.message = matchGroups[0];
         }
@@ -108,7 +110,9 @@ export class CdkErrorMapper {
       classification: 'ERROR',
     },
     {
-      errorRegex: /(SyntaxError|ReferenceError|TypeError):((?:.|\n)*?at .*)/,
+      errorRegex: new RegExp(
+        `(SyntaxError|ReferenceError|TypeError):((?:.|${EOL})*?at .*)`
+      ),
       humanReadableErrorMessage:
         'Unable to build the Amplify backend definition.',
       resolutionMessage:
@@ -134,8 +138,9 @@ export class CdkErrorMapper {
       classification: 'ERROR',
     },
     {
-      errorRegex:
-        /\[ERR_MODULE_NOT_FOUND\]:(.*)\n|Error: Cannot find module (.*)/,
+      errorRegex: new RegExp(
+        `\\[ERR_MODULE_NOT_FOUND\\]:(.*)${EOL}|Error: Cannot find module (.*)`
+      ),
       humanReadableErrorMessage: 'Cannot find module',
       resolutionMessage:
         'Check your backend definition in the `amplify` folder for missing file or package imports. Try installing them with your package manager.',
@@ -154,11 +159,21 @@ export class CdkErrorMapper {
     },
     {
       // Also extracts the first line in the stack where the error happened
-      errorRegex: /\[esbuild Error\]: ((?:.|\n)*?at .*)/,
+      errorRegex: new RegExp(`\\[esbuild Error\\]: ((?:.|${EOL})*?at .*)`),
       humanReadableErrorMessage:
         'Unable to build the Amplify backend definition.',
       resolutionMessage:
         'Check your backend definition in the `amplify` folder for syntax and type errors.',
+      errorName: 'ESBuildError',
+      classification: 'ERROR',
+    },
+    {
+      errorRegex: new RegExp(
+        `\\[TransformError\\]: Transform failed with .* error:${EOL}(?<esBuildErrorMessage>.*)`
+      ),
+      humanReadableErrorMessage: '{esBuildErrorMessage}',
+      resolutionMessage:
+        'Fix the above mentioned type or syntax error in your backend definition.',
       errorName: 'ESBuildError',
       classification: 'ERROR',
     },
@@ -203,7 +218,7 @@ export class CdkErrorMapper {
     {
       // Error: .* is printed to stderr during cdk synth
       // Also extracts the first line in the stack where the error happened
-      errorRegex: /^Error: (.*\n.*at.*)/m,
+      errorRegex: new RegExp(`^Error: (.*${EOL}.*at.*)`, 'm'),
       humanReadableErrorMessage:
         'Unable to build the Amplify backend definition.',
       resolutionMessage:
@@ -223,8 +238,8 @@ export class CdkErrorMapper {
     {
       // We capture the parameter name to show relevant error message
       errorRegex:
-        /Failed to retrieve backend secret (.*) for.*ParameterNotFound/,
-      humanReadableErrorMessage: `The secret ${this.placeHolder} specified in the backend does not exist.`,
+        /Failed to retrieve backend secret (?<secretName>.*) for.*ParameterNotFound/,
+      humanReadableErrorMessage: `The secret {secretName} specified in the backend does not exist.`,
       resolutionMessage: `Create secrets using the command ${this.formatter.normalizeAmpxCommand(
         'sandbox secret set'
       )}. For more information, see https://docs.amplify.aws/gen2/deploy-and-host/sandbox-environments/features/#set-secrets`,
@@ -233,7 +248,7 @@ export class CdkErrorMapper {
     },
     {
       // Note that the order matters, this should be the last as it captures generic CFN error
-      errorRegex: /❌ Deployment failed: (.*)\n/,
+      errorRegex: new RegExp(`❌ Deployment failed: (.*)${EOL}`),
       humanReadableErrorMessage: 'The CloudFormation deployment has failed.',
       resolutionMessage:
         'Find more information in the CloudFormation AWS Console for this stack.',