+++ packages/deployed-backend-client/src/backend_output_client_factory.ts
@@ -8,6 +8,9 @@ import { AWSClientProvider } from '@aws-amplify/plugin-types';
 
 export enum BackendOutputClientErrorType {
   METADATA_RETRIEVAL_ERROR = 'MetadataRetrievalError',
+  NO_OUTPUTS_FOUND = 'NoOutputsFound',
+  DEPLOYMENT_IN_PROGRESS = 'DeploymentInProgress',
+  NO_STACK_FOUND = 'NoStackFound',
 }
 /**
  * Error type for BackendOutputClientError