+++ packages/integration-tests/src/test-e2e/backend_output.test.ts
@@ -21,6 +21,8 @@ import { S3Client } from '@aws-sdk/client-s3';
 import { IAMClient } from '@aws-sdk/client-iam';
 import { DeployedResourcesFinder } from '../find_deployed_resource.js';
 import { DataStorageAuthWithTriggerTestProjectCreator } from '../test-project-setup/data_storage_auth_with_triggers.js';
+import { SQSClient } from '@aws-sdk/client-sqs';
+import { setupDeployedBackendClient } from '../test-project-setup/setup_deployed_backend_client.js';
 
 // Different root test dir to avoid race conditions with e2e deployment tests
 const rootTestDir = fileURLToPath(
@@ -37,6 +39,7 @@ void describe(
     const lambdaClient = new LambdaClient(e2eToolingClientConfig);
     const s3Client = new S3Client(e2eToolingClientConfig);
     const iamClient = new IAMClient(e2eToolingClientConfig);
+    const sqsClient = new SQSClient(e2eToolingClientConfig);
     const resourceFinder = new DeployedResourcesFinder(cfnClient);
     const dataStorageAuthWithTriggerTestProjectCreator =
       new DataStorageAuthWithTriggerTestProjectCreator(
@@ -46,6 +49,7 @@ void describe(
         lambdaClient,
         s3Client,
         iamClient,
+        sqsClient,
         resourceFinder
       );
 
@@ -59,6 +63,7 @@ void describe(
         await dataStorageAuthWithTriggerTestProjectCreator.createProject(
           rootTestDir
         );
+      await setupDeployedBackendClient(testProject.projectDirPath);
       testBranch = await amplifyAppPool.createTestBranch();
       branchBackendIdentifier = {
         namespace: testBranch.appId,
@@ -75,6 +80,7 @@ void describe(
       const sharedSecretsEnv = {
         [amplifySharedSecretNameKey]: createAmplifySharedSecretName(),
       };
+
       await testProject.deploy(branchBackendIdentifier, sharedSecretsEnv);
       await testProject.assertPostDeployment(branchBackendIdentifier);