+++ packages/platform-core/src/errors/amplify_error.test.ts
@@ -148,3 +148,35 @@ and some after the error message
     assert.deepStrictEqual(actual?.resolution, 'test resolution');
   });
 });
+
+void describe('AmplifyError.fromError', async () => {
+  void it('wraps Yargs validation errors in AmplifyUserError', () => {
+    const yargsErrors = [
+      new Error('Unknown command: asd'),
+      new Error('Unknown arguments: a,b,c'),
+      new Error('Missing required argument: d'),
+      new Error('Did you mean sandbox'),
+      new Error('Not enough non-option arguments: got 2, need at least 4'),
+      new Error('Invalid values: Arguments: a, Given: n, Choices: [b, c, d]'),
+      new Error('Arguments a and b are mutually exclusive'),
+      new Error('Missing dependent arguments: branch -> appId'),
+      new Error('Implications failed:\n app-id -> branch'),
+    ];
+    yargsErrors.forEach((error) => {
+      const actual = AmplifyError.fromError(error);
+      assert.ok(
+        actual instanceof AmplifyError &&
+          actual.name === 'InvalidCommandInputError',
+        `Failed the test for error ${error.message}`
+      );
+    });
+  });
+  void it('wraps getaddrinfo ENOTFOUND errors in AmplifyUserError', () => {
+    const error = new Error('getaddrinfo ENOTFOUND some-domain.com');
+    const actual = AmplifyError.fromError(error);
+    assert.ok(
+      actual instanceof AmplifyError && actual.name === 'DomainNotFoundError',
+      `Failed the test for error ${error.message}`
+    );
+  });
+});